Make order

+++ app/src/main/java/edu/uoc/easyorderfront/data/menu/MenuResponse.kt
@@ -36,7 +36,7 @@ data class DishDTO(
     val uid: String?,
     val name: String? = null,
     val description: String? = null,
-    val price: Double = 0.0,
+    val price: Float = 0F,
     val aliments: MutableList<String>? = ArrayList(),
     val calories: Int? = 0
 ) {
